Move tests to PSR-4 structure
