+++ tests/Rules/data/CorrectModelInstantiation.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Rules\Data;
+namespace Tests\Rules\data;
 
 use App\Group;
 use App\User;
