+++ tests/Rules/data/LaravelJobWithoutConstructor.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Rules\data;
+
+use Illuminate\Foundation\Bus\Dispatchable;
+
+class LaravelJobWithoutConstructor
+{
+    use Dispatchable;
+}
