+++ tests/Rules/data/UnnecessaryCollectionCallsQuery.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Rules\Data;
+namespace Tests\Rules\data;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
