+++ tests/Rules/data/event-dispatch.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Rules\Data;
+namespace Tests\Rules\data;
 
 LaravelEvent::dispatch();
