+++ tests/Rules/data/model-property-model-factory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ModelPropertyModelFactory;
+namespace Tests\Rules\data;
 
 function foo(): void
 {
