+++ tests/Unit/ModelFactoryDynamicStaticMethodReturnTypeExtensionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Unit;
+namespace Tests\Unit;
 
 use NunoMaduro\Larastan\ReturnTypes\ModelFactoryDynamicStaticMethodReturnTypeExtension;
 use NunoMaduro\Larastan\Types\Factory\ModelFactoryType;
