+++ tests/Unit/data/additional_migrations/2021_02_01_000000_create_teams_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\AdditionalMigrations;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateTeamsTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -22,4 +20,4 @@ class CreateTeamsTable extends Migration
             $table->timestamps();
         });
     }
-}
+};
