+++ tests/Unit/data/complex_migrations/2020_01_30_000000_create_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\ComplexMigrations;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -22,4 +20,4 @@ class CreateUsersTable extends Migration
             $table->timestamps();
         });
     }
-}
+};
