+++ tests/Unit/data/complex_migrations/2020_01_31_000000_alter_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\ComplexMigrations;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AlterUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -45,4 +43,4 @@ class AlterUsersTable extends Migration
             $table->integer('active');
         });
     }
-}
+};
