+++ tests/Unit/data/migration_with_schema_connection/2020_01_30_000000_create_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\MigrationWithSchemaConnection;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class() extends Migration
 {
     public function up(): void
     {
@@ -19,4 +17,4 @@ class CreateUsersTable extends Migration
             $table->timestamps();
         });
     }
-}
+};
