+++ tests/Unit/data/migrations_using_after_method/2020_01_31_000000_add_to_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\MigrationsUsingAfterMethod;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AlterUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -21,4 +19,4 @@ class AlterUsersTable extends Migration
             });
         });
     }
-}
+};
