+++ tests/Unit/data/migrations_using_soft_deletes/2020_01_30_000000_create_users_table_with_soft_deletes.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\MigrationsUsingSoftDeletes;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -23,4 +21,4 @@ class CreateUsersTable extends Migration
             $table->softDeletes();
         });
     }
-}
+};
