+++ tests/Unit/data/migrations_using_soft_deletes_tz/2020_01_31_000000_alter_users_table_to_use_soft_deletes_tz.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\MigrationsUsingSoftDeletesTz;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -23,4 +21,4 @@ class CreateUsersTable extends Migration
             $table->softDeletesTz();
         });
     }
-}
+};
