+++ tests/Unit/data/migrations_with_different_methods/2020_01_30_000000_create_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Migrations;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -32,4 +30,4 @@ class CreateUsersTable extends Migration
             $table->timestamps();
         });
     }
-}
+};
