+++ tests/Unit/data/rename_migrations/2021_12_08_000000_rename_users_table.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\RenameMigrations;
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class RenameUsersTable extends Migration
+return new class() extends Migration
 {
     /**
      * Run the migrations.
@@ -26,4 +24,4 @@ class RenameUsersTable extends Migration
             $table->rename('accounts');
         });
     }
-}
+};
